+++ src/NetCore/Web/Api/wwwroot/scripts/api/webgis.l10n.de.js
@@ -24,6 +24,8 @@
     "year": "Jahr",
     "month": "Monat",
     "day": "Tag",
+    "point-in-time": "Zeitpunkt",
+    "span-of-time": "Zeitspanne",
 
     "query-results": "Abfrageergebnisse",
     "search-results": "Suchergebnisse",